backend/pyproject.toml
[tool.poetry]
name = "helenus-ai"
version = "0.0.1"
description = "An AI automated trading agent"
authors = ["John Peterson <john.peterson@coinbase.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
cdp-langchain = "^0.0.13"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
websockets = "^12.0"
pydantic = "^2.5.3"
python-dotenv = "^1.0.0"
web3 = "^6.14.0"
aiohttp = "^3.9.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
